+++ android-sample/src/main/java/io/noties/tumbleweed/sample/MainActivity.java
@@ -4,7 +4,17 @@ import android.app.Activity;
 import android.app.Fragment;
 import android.app.FragmentManager;
 import android.os.Bundle;
+import android.view.View;
 
+import io.noties.tumbleweed.Timeline;
+import io.noties.tumbleweed.Tween;
+import io.noties.tumbleweed.android.ViewTweenManager;
+import io.noties.tumbleweed.android.types.Alpha;
+import io.noties.tumbleweed.android.types.Pivot;
+import io.noties.tumbleweed.android.types.Rotation;
+import io.noties.tumbleweed.android.types.Scale;
+import io.noties.tumbleweed.android.types.Translation;
+import io.noties.tumbleweed.android.utils.ViewUtils;
 import io.noties.tumbleweed.sample.anim.BaseAnimationFragment;
 import io.noties.tumbleweed.sample.drawable.DrawableFragment;
 import io.noties.tumbleweed.sample.easing.EasingFragment;
@@ -24,6 +34,22 @@ public class MainActivity extends Activity implements MenuFragment.Callbacks {
             manager.beginTransaction()
                     .replace(R.id.container, MenuFragment.newInstance(false), TAG_MENU)
                     .commit();
+        }
+
+        if (true) {
+            final ViewTweenManager tweenManager = ViewTweenManager.get(findViewById(R.id.container));
+            final View view = tweenManager.view();
+
+            ViewUtils.whenReady(view, $ -> {
+                Timeline.createParallel(1.0F)
+                        .push(Tween.to(view, Alpha.VIEW).target(0.0F))
+                        .push(Tween.to(view, Scale.XY).target(0.0F, 0.0F))
+                        .push(Tween.to(view, Rotation.I).target(270))
+                        .push(Tween.to(view, Pivot.XY).target(0, 0))
+                        .push(Tween.to(view, Translation.XY).target($.getWidth(), $.getHeight()))
+                        .repeatYoyo(-1, 0)
+                        .start(tweenManager);
+            });
         }
     }