+++ tumbleweed-android-kt/src/main/java/io/noties/tumbleweed/android/kt/ArgEx.kt
@@ -5,8 +5,8 @@ package io.noties.tumbleweed.android.kt
 
 import io.noties.tumbleweed.android.types.Argb
 
-public fun Int.toArgbArray(): FloatArray = Argb.toArray(this)
+fun Int.toArgbArray(): FloatArray = Argb.toArray(this)
 
-public fun Int.toArgbArray(array: FloatArray): FloatArray = Argb.toArray(this, array)
+fun Int.toArgbArray(array: FloatArray): FloatArray = Argb.toArray(this, array)
 
-public fun FloatArray.toColor() = Argb.fromArray(this)
+fun FloatArray.toColor() = Argb.fromArray(this)