+++ tumbleweed-android-kt/src/main/java/io/noties/tumbleweed/android/kt/DrawableEx.kt
@@ -9,20 +9,30 @@ import io.noties.tumbleweed.TweenDef
 import io.noties.tumbleweed.TweenType
 import io.noties.tumbleweed.android.DrawableTweenManager
 
-public fun Drawable.tweenManager() = DrawableTweenManager.create(this)
+fun Drawable.tweenManager() = DrawableTweenManager.create(this)
 
-public fun Drawable.tweenManager(frame: Float) = DrawableTweenManager.create(this, frame)
+fun Drawable.tweenManager(frame: Float) = DrawableTweenManager.create(this, frame)
 
 /**
  * Create a [TweenDef] with caller Drawable as the target. Please note that resulting [TweenDef]
  * doesn\'t have duration specified and it must be set explicitly via [TweenDef.duration] method call.
  */
-public fun Drawable.tween(type: TweenType<Drawable>): TweenDef<Drawable> =
+fun Drawable.tween(type: TweenType<Drawable>): TweenDef<Drawable> =
         Tween.to(this, type)
 
-public fun Drawable.tween(type: TweenType<Drawable>, duration: Float): TweenDef<Drawable> =
+fun Drawable.tween(type: TweenType<Drawable>, duration: Float): TweenDef<Drawable> =
         Tween.to(this, type, duration)
 
-public fun Drawable.applyIntrinsicBounds() {
+fun Drawable.applyIntrinsicBounds() {
     setBounds(0, 0, intrinsicWidth, intrinsicHeight)
+}
+
+/**
+ * Will set intrinsic bounds if only current bounds are empty
+ * @since 2.1.0-SNAPSHOT
+ */
+fun Drawable.applyIntrinsicBoundsIfEmpty() {
+    if (bounds.isEmpty) {
+        applyIntrinsicBounds()
+    }
 }