+++ tumbleweed-android-kt/src/main/java/io/noties/tumbleweed/android/kt/DurationEx.kt
@@ -3,6 +3,6 @@
  */
 package io.noties.tumbleweed.android.kt
 
-public fun Int.toFloatSeconds(): Float = this / 1000.0F
+fun Int.toFloatSeconds(): Float = this / 1000.0F
 
-public fun Long.toFloatSeconds(): Float = this / 1000.0F
+fun Long.toFloatSeconds(): Float = this / 1000.0F