+++ tumbleweed/src/main/java/io/noties/tumbleweed/Timeline.java
@@ -55,6 +55,18 @@ public final class Timeline extends BaseTween {
     }
 
     /**
+     * @param defaultTweenDuration default duration for <em>actionable</em> tweens created with
+     *                             {@link Tween#to(Object, TweenType)} and {@link Tween#from(Object, TweenType)}.
+     *                             Tweens with duration specified and Timelines are ignored and will use own duration.
+     * @see #createParallel(float)
+     * @since 2.1.0-SNAPSHOT
+     */
+    @NonNull
+    public static TimelineDef createSequence(float defaultTweenDuration) {
+        return new TimelineDefImpl(TimelineDefImpl.Mode.SEQUENCE, defaultTweenDuration);
+    }
+
+    /**
      * Creates a new timeline with a 'parallel' behavior. Its children will be
      * triggered all at once.
      */
@@ -62,6 +74,18 @@ public final class Timeline extends BaseTween {
     @NonNull
     public static TimelineDef createParallel() {
         return new TimelineDefImpl(TimelineDefImpl.Mode.PARALLEL);
+    }
+
+    /**
+     * @param defaultTweenDuration default duration for <em>actionable</em> tweens created with
+     *                             {@link Tween#to(Object, TweenType)} and {@link Tween#from(Object, TweenType)}.
+     *                             Tweens with duration specified and Timelines are ignored and will use own duration.
+     * @see #createSequence(float)
+     * @since 2.1.0-SNAPSHOT
+     */
+    @NonNull
+    public static TimelineDef createParallel(float defaultTweenDuration) {
+        return new TimelineDefImpl(TimelineDefImpl.Mode.PARALLEL, defaultTweenDuration);
     }
 
     private final float duration;